Start on splitting up EventManager
